First commit MaDaDevAPI on the GitHub
This repository contains API intended for using on the site MaDaDev and
another developers. If you want to change this API or make this API better
please contact the author: infinityhorizon@mail.ru

Start the API and Databases

The beginning of writing this API was on January 31, 2021. Make two
database, first for main part of Django auto-created schemas, and second for
another tables, calculated in order not to get confused in the tables.

Apps in API

Main app is the most big at this time, but multiplayer app
planned to be wider due to its original purpose.