Temporarily removed OAuth
This was necessary because using JWT and OAuth at the
same time is not advisable.

Setting permissions

Since now the use of logging into the site is not protected
by a token, now there are access rights. Fine-tuning of
rights is required to use all possible options of this API.

Small updates

Improved password recovery, as well as email confirmation.